requirements.txt
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.3
conda==4.8.4
conda-build==3.18.8
conda-package-handling==1.7.0
conda-verify==3.4.2
contextlib2==0.6.0.post1
cryptography==2.9.2
cvxopt==1.2.5
decorator==4.4.2
dgl-cu90==0.4.3.post2
dgllife==0.2.4
numpy==1.19.1
ply==3.11
plyfile==0.7.1
QtPy==1.9.0
quadprog==0.1.6
scikit-image==0.16.2
scikit-learn==0.23.2
scipy==1.1.0
torch==1.1.0
torch-cluster==1.4.5
torch-geometric==1.3.2
torch-scatter==1.4.0
torch-sparse==0.4.3
torch-spline-conv==1.1.1
torchfile==0.1.0
torchvision==0.3.0
tornado==6.0.4

